requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
matplotlib
seaborn
crewai
python-dotenv
litellm
tenacity
